core: api: Separate gossip and external APIs for discoverability
