+++ core/src/external_api/mod.rs
@@ -3,12 +3,7 @@
 
 use serde::{Deserialize, Serialize};
 
-pub mod cluster_management;
-pub mod gossip;
-pub mod handshake;
-pub mod heartbeat;
 pub mod http;
-pub mod orderbook_management;
 pub mod websocket;
 
 /// An empty request/response type
