+++ core/src/gossip_api/gossip.rs
