+++ core/src/gossip_api/mod.rs
@@ -1,0 +1,7 @@
+//! Defines API types for gossip within the p2p network
+
+pub mod cluster_management;
+pub mod gossip;
+pub mod handshake;
+pub mod heartbeat;
+pub mod orderbook_management;
